+++ src/main/java/com/supermartijn642/chunkloaders/ChunkLoaderBlock.java
@@ -34,7 +34,6 @@ import net.minecraft.world.level.material.MapColor;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Consumer;
 
@@ -118,7 +117,7 @@ public class ChunkLoaderBlock extends BaseBlock implements EntityHoldingBlock, S
     }
 
     @Override
-    protected void appendItemInformation(ItemStack stack, @Nullable BlockGetter level, Consumer<Component> info, boolean advanced){
+    protected void appendItemInformation(ItemStack stack, Consumer<Component> info, boolean advanced){
         if(this.type.getGridSize() == 1)
             info.accept(TextComponents.translation("chunkloaders.chunk_loader.info.single").color(ChatFormatting.AQUA).get());
         else