+++ changelog.md
@@ -1,2 +1,2 @@
 ### Chunk Loaders 1.2.8
-- Initial release of Chunk Loaders for NeoForge
+- Initial release of Chunk Loaders for Minecraft 1.20.5 & 1.20.6